SegundoParcial/backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pydantic[email]
python-multipart
# Seguridad
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
google-auth
requests